requirements.txt
Flask==3.0.3
opencv-python-headless==4.10.0.84
sounddevice==0.4.7
soundfile==0.12.1
PyYAML==6.0.2
